architecture_1/requirements.txt
# asgiref             3.8.1
# blinker             1.8.2
# click               8.1.8
# colorama            0.4.6
# Django              5.2
# django-cors-headers 4.7.0
# django-filter       25.1
# django-rest         0.8.7
# djangorestframework 3.16.0
# exceptiongroup      1.2.2
# Faker               35.2.2
# Flask               3.0.3
# Flask-Login         0.6.3
# gunicorn            23.0.0
# importlib_metadata  8.5.0
# iniconfig           2.0.0
# itsdangerous        2.2.0
# Jinja2              3.1.6
# Markdown            3.8
# MarkupSafe          2.1.5
# packaging           24.2
# pip                 24.3.1
# pluggy              1.5.0
# psycopg             3.2.6
# psycopg-binary      3.2.6
# psycopg-pool        3.2.6
# pytest              8.3.5
# pytest-mock         3.14.0
# python-dateutil     2.9.0.post0
# python-dotenv       1.0.1
# six                 1.17.0
# sqlparse            0.5.3
# tomli               2.2.1
# typing_extensions   4.12.2
# tzdata              2025.2
# Werkzeug            3.0.6
# zipp                3.20.2
djangorestframework
markdown
django-filter
psycopg2-binary
Faker
django-cors-headers
Django>=4.2
gunicorn
